Change CTCoreMessage.render and CTMIME.render to return NSData, not NSString, and change CTSMTP.setData to take NSData to match.
This avoids interpreting the bytes that mailmime_write_mem gives us
as UTF-8, only to write them back out as UTF-8 when sending the
message.

As part of this, mention that the render methods return something
that needs to be released using mmap_string_unref.  This avoids
a small leak each time this method is called, because the render
call was using mmap_string_ref and then passing the bytes to
NSString.initWithBytes.  This was leaking the hashtable entry that
mmap_string_ref makes.