Fix appendMessage:flags:extensionFlags:
This was sending the bytes of the mailbox path, but passing the length of
the message to be appended.  This makes no sense at all, and if it doesn't
just crash it'll append an email containing garbage.

This came in from fbd8a3521d on pitiphong-p/appenduid_support, which has
never actually been committed upstream.  The similar-looking fix down in
appendMessage:appendedUID:flags:extensionFlags: is correct, it's just this
one that is garbage.