Fix a memory leak when sending an email with a message-part attachment.
The ownership for the mailimf_fields given to CTMIME_MessagePart was
screwed up.  There was a hack in CTMIME.renderData to work around a
double-free, but that was just masking the real problem, which is that
both the CTMIME_MessagePart and the mailmime struct returned by buildMIMEStruct
wanted to own the mailimf_fields.  This wasn't a problem when there was
only a single message-part, thanks to the aforementioned hack, but this
meant that there was a memory leak if the message-part was not at the top
level.  There was also a "TODO uh oh" which was acknowledging this problem.

Fix this by changing CTMIME_MessagePart so that it transfers ownership
of the mailimf_fields correctly, and frees it in dealloc if it owns it at
that point.

Also, add a comment in a number of places noting that the CTCoreMessage and
the underlying CTMIME instances will be mutated whenever a renderData call
is made, because the mailimf_fields is being transferred over.